+++ game_patch/rf/input.h
@@ -118,6 +118,7 @@ namespace rf
     };
 
     static auto& mouse_get_pos = addr_as_ref<int(int &x, int &y, int &z)>(0x0051E450);
+    static auto& mouse_get_delta = addr_as_ref<void(int& dx, int& dy, int& dz)>(0x0051E630);
     static auto& mouse_was_button_pressed = addr_as_ref<int(int btn_idx)>(0x0051E5D0);
     static auto& mouse_button_is_down = addr_as_ref<bool(int btn_idx)>(0x0051E530);
     static auto& mouse_set_visible = addr_as_ref<void(bool visible)>(0x0051E680);